pyproject.toml
[tool.poetry]
name = "parsee-datasets"
version = "0.1.2"
description = ""
authors = ["Parsee.ai <info@parsee.ai>"]
homepage = "https://parsee.ai"
packages = [{include = "parsee"}]

[tool.poetry.dependencies]
python = ">=3.9,<3.11"
pandas = "^1.4.3"
parsee-core = "^0.1.4.11"
pyarrow = "^15.0.1"
jupyterlab = "^4.1.5"
matplotlib = "^3.8.3"


[tool.poetry.dev-dependencies]

[tool.poetry.group.extras.dependencies]
langchain = "^0.1.11"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
